CODE/Protego/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
twilio==9.4.2
asyncpg==0.30.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for compatibility with passlib
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
slowapi==0.1.9
loguru==0.7.3
sentry-sdk[fastapi]==2.21.0
redis==5.0.1
hiredis==2.3.2
gunicorn==21.2.0
deepgram-sdk==5.3.1
elevenlabs==2.30.0
boto3==1.34.19
reportlab==4.0.9
matplotlib==3.8.2
pandas==2.1.4
